Separate nginx templates and includes check
Previously the templates were patched if the includes were unpatched,
this happened in the case when the templates were previously patched but
the nginx restart was rejected or failed. This resulted in the templates
being patched multiple times.
Now we check for patched templates first, then check for unpatched includes
and restart nginx if necessary.