Use topmost certificate in chain to find CA cert. Fixes #129
The script assumed chain.pem had only one intermediate certificate in
it. This used to work until letsencrypt started using 2 intermediates
which were concatenated in chain.pem. Since openssl x509 parses only the
first one in the file, but the topmost one is actually the last in the
file, the script was appending the wrong CA cert to the full chain that
zmcertmgr expects.

Using bash string manipulation to extract the last cerftificate in the
file (topmost in the chain) fixes this.