Rewrite logic for finding CA certificate from chain
The old method was bad as it only picked one certificate from the chain
and tried to find its issuer. This didn't work if it happened to pick
a chain that was invalid. For example, Letsencrypt has two chains, where
one is invalid (signed by "DST Root CA X3" which is now expired) and one
is valid (signed by "ISRG Root X1"). We can't just pick the first, last,
or n-th certificate from the chain to find a valid chain. We have to
iterate through each certificate in the chain, check if it's valid using
"openssl verify" and use the CA of the first valid certificate.

In the process of fixing this I also found that Zimbra's zmcertmgr will
fail to validate a chain that has the CA at the end, it wants to have
the CA as the first certificate. It passes the chain file to "openssl
verify" as "-CAfile".

We can also use "openssl x509 -issuer -nameopt sep_multiline,utf8" to
make it easier to parse the CN and can replace a complicated "sed"
script with a simple "grep".

Fixes #140