Syntax: disable pathname expansion while looping over unquoted variable
RENEWED_DOMAINS is passed as an environment variable from Certbot.
It's a space-delimited list string. Looping over it with for is safe if
pathname expansion is disabled, then only word splitting on IFS remains,
which is not a problem, as the list should never contain tabs or
newlines, just spaces (IFS default is '<space><tab><newline>').
Other methods of looping would be more complicated and are not needed in
this case.