Remove certbot-auto version detection bootstrapping hack
In certbot version detection code certbot-auto needed a workaround in
case it's its first invocation after installation and it needed to
bootstrap, needing an interactive terminal. Since certbot-auto has
been dead for a long time, it's safe to assume noone is still using it,
if they are, it's safe to assume it won't need to bootstrap.